ppo_diffusion/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.16.0
peft>=0.6.0
numpy>=1.21.0
Pillow>=8.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0